refactor: Restructure the fetchers code to not use more memory than needed
Also fixes: #92
